Melhorias + mocks pra testes

+++ src/main/java/com/softdesign/livetest/api/produto/CreateProdutoRequest.java
@@ -2,10 +2,11 @@ package com.softdesign.livetest.api.produto;
 
 import com.softdesign.livetest.domain.produto.Produto;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public record CreateProdutoRequest(String nome, String descricao, Double valor) {
+public record CreateProdutoRequest(String nome, String descricao, BigDecimal valor) {
 
     public Produto toProduto() {
         return new Produto(null, this.nome(), this.descricao(), this.valor());
