+++ src/main/java/com/softdesign/livetest/api/produto/ProdutoDTO.java
@@ -7,10 +7,11 @@ import com.softdesign.livetest.domain.venda.ItemVenda;
 import jakarta.validation.constraints.NotBlank;
 import org.springframework.data.annotation.Id;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public record ProdutoDTO(@NotBlank(message = "Id do produto é obrigatório")String id, @NotBlank(message = "Nome do produto é obrigatório")String nome, String descricao, Double valor) {
+public record ProdutoDTO(@NotBlank(message = "Id do produto é obrigatório")String id, @NotBlank(message = "Nome do produto é obrigatório")String nome, String descricao, BigDecimal valor) {
 
 
     public Produto toProduto() {
