+++ src/main/java/com/softdesign/livetest/api/produto/UpdateProdutoRequest.java
@@ -2,7 +2,9 @@ package com.softdesign.livetest.api.produto;
 
 import com.softdesign.livetest.domain.produto.Produto;
 
-public record UpdateProdutoRequest(String nome, String descricao, Double valor) {
+import java.math.BigDecimal;
+
+public record UpdateProdutoRequest(String nome, String descricao, BigDecimal valor) {
 
     public Produto toProduto(String id) {
         return new Produto(id, this.nome(), this.descricao(), this.valor());
