+++ src/main/java/com/softdesign/livetest/api/venda/ItemVendaDTO.java
@@ -5,10 +5,11 @@ import com.softdesign.livetest.domain.produto.Produto;
 import com.softdesign.livetest.domain.venda.ItemVenda;
 import com.softdesign.livetest.domain.venda.Venda;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public record ItemVendaDTO(String produtoId, Integer quantidade, Double valorTotal) {
+public record ItemVendaDTO(String produtoId, Integer quantidade, BigDecimal valorTotal) {
 
     public static ItemVendaDTO from(ItemVenda itemVenda) {
         return new ItemVendaDTO(itemVenda.produtoId(), itemVenda.quantidade(), itemVenda.valorTotal());
