+++ src/main/java/com/softdesign/livetest/domain/produto/Produto.java
@@ -2,5 +2,7 @@ package com.softdesign.livetest.domain.produto;
 
 import org.springframework.data.annotation.Id;
 
-public record Produto(@Id String id, String nome, String descricao, Double valor) {
+import java.math.BigDecimal;
+
+public record Produto(@Id String id, String nome, String descricao, BigDecimal valor) {
 }
