+++ src/main/java/com/softdesign/livetest/domain/venda/ItemVenda.java
@@ -1,12 +1,14 @@
 package com.softdesign.livetest.domain.venda;
 
-public record ItemVenda(String produtoId, Integer quantidade, Double valorTotal) {
+import java.math.BigDecimal;
+
+public record ItemVenda(String produtoId, Integer quantidade, BigDecimal valorTotal) {
 
     public ItemVenda withQuantidade(Integer quantidade) {
         return new ItemVenda(this.produtoId, quantidade, this.valorTotal);
     }
 
-    public ItemVenda withValorTotal(Double valorTotal) {
+    public ItemVenda withValorTotal(BigDecimal valorTotal) {
         return new ItemVenda(this.produtoId, this.quantidade, valorTotal);
     }
