+++ src/main/java/com/softdesign/livetest/domain/venda/Venda.java
@@ -2,11 +2,12 @@ package com.softdesign.livetest.domain.venda;
 
 import org.springframework.data.annotation.Id;
 
+import java.math.BigDecimal;
 import java.util.List;
 
-public record Venda (@Id String id, String clienteId, Double valorTotal, List<ItemVenda> items) {
+public record Venda (@Id String id, String clienteId, BigDecimal valorTotal, List<ItemVenda> items) {
 
-    public Venda withValorTotal (Double valorTotal) {
+    public Venda withValorTotal (BigDecimal valorTotal) {
         return new Venda(this.id, this.clienteId, valorTotal, this.items);
     }
