+++ src/main/java/com/softdesign/livetest/migration/mongo/DatabaseChangeLog0002.java
@@ -7,14 +7,16 @@ import io.mongock.api.annotations.ChangeUnit;
 import io.mongock.api.annotations.Execution;
 import io.mongock.api.annotations.RollbackExecution;
 
+import java.math.BigDecimal;
+
 @ChangeUnit(id = "0002", order = "2", author = "ricardo.camelo")
 public class DatabaseChangeLog0002 {
 
     @Execution
     public void execution(ProdutoMongoRepository produtoMongoRepository) {
-        produtoMongoRepository.save(new Produto(null, "Sacolé", "Sabor laranja", 3.00d));
-        produtoMongoRepository.save(new Produto(null, "Parafuso", "Tamanho médio", 1.00d));
-        produtoMongoRepository.save(new Produto(null, "Ovos", "Duzia de ovos", 10.00d));
+        produtoMongoRepository.save(new Produto(null, "Sacolé", "Sabor laranja", BigDecimal.valueOf(3.00d)));
+        produtoMongoRepository.save(new Produto(null, "Parafuso", "Tamanho médio", BigDecimal.valueOf(1.00d)));
+        produtoMongoRepository.save(new Produto(null, "Ovos", "Duzia de ovos", BigDecimal.valueOf(10.00d)));
     }
 
     @RollbackExecution
